+++ src/hooks/common/inputsUserInfo/useCheckEmail.ts
@@ -1,25 +1,26 @@
 import { useRef, useState } from 'react';
-import { isEmptyString } from '../utils/validator';
-import { ERROR } from '../constants/message';
+import { isEmptyString } from '../../../utils/validator';
+import { ERROR } from '../../../constants/message';
 
-export const useCheckNickname = (value: string) => {
+export const useCheckEmail = (value: string) => {
   const inputRef = useRef<HTMLInputElement | null>(null);
   const [errorMessage, setErrorMessage] = useState<string | undefined>(
     undefined,
   );
   const [isError, setIsError] = useState<boolean>(false);
 
-  function isValidNickname(nickname: string): boolean {
-    return nickname.length >= 2 && nickname.length <= 10;
-  }
+  const isValidEmailFormat = (email: string): boolean => {
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    return emailRegex.test(email);
+  };
 
   const handleBlur = () => {
     if (isEmptyString(value)) {
       setIsError(true);
-      setErrorMessage(ERROR.NICKNAME.EMPTY);
-    } else if (!isValidNickname(value)) {
+      setErrorMessage(ERROR.EMAIL.EMPTY);
+    } else if (!isValidEmailFormat(value)) {
       setIsError(true);
-      setErrorMessage(ERROR.NICKNAME.FORM);
+      setErrorMessage(ERROR.EMAIL.FORM);
     } else {
       setIsError(false);
       setErrorMessage(undefined);
