+++ src/hooks/common/inputsUserInfo/useCheckNickname.ts
@@ -1,30 +1,25 @@
 import { useRef, useState } from 'react';
-import { isEmptyString } from '../utils/validator';
-import { ERROR } from '../constants/message';
+import { isEmptyString } from '../../../utils/validator';
+import { ERROR } from '../../../constants/message';
 
-interface CheckPwChkProps {
-  value: string;
-  pw: string;
-}
-
-export const useCheckPasswordCheck = ({ value, pw }: CheckPwChkProps) => {
+export const useCheckNickname = (value: string) => {
   const inputRef = useRef<HTMLInputElement | null>(null);
   const [errorMessage, setErrorMessage] = useState<string | undefined>(
     undefined,
   );
   const [isError, setIsError] = useState<boolean>(false);
 
-  const isValidPwChkMatch = (pwChk: string): boolean => {
-    return pwChk === pw;
-  };
+  function isValidNickname(nickname: string): boolean {
+    return nickname.length >= 2 && nickname.length <= 10;
+  }
 
   const handleBlur = () => {
     if (isEmptyString(value)) {
       setIsError(true);
-      setErrorMessage(ERROR.PW.EMPTY);
-    } else if (!isValidPwChkMatch(value)) {
+      setErrorMessage(ERROR.NICKNAME.EMPTY);
+    } else if (!isValidNickname(value)) {
       setIsError(true);
-      setErrorMessage(ERROR.PW.NOT_MATCH);
+      setErrorMessage(ERROR.NICKNAME.FORM);
     } else {
       setIsError(false);
       setErrorMessage(undefined);
